# airflow/api_connexion/openapi/v1.yaml
...
openapi: 3.0.3
info:
  title: "Airflow API (Stable)"
  description: |
    # Overview

    To facilitate management, Apache Airflow supports a range of REST API endpoints across its
    objects.
    This section provides an overview of the API design, methods, and supported use cases.

    Most of the endpoints accept `JSON` as input and return `JSON` responses.
    This means that you must usually add the following headers to your request:
    ```
    Content-type: application/json
    Accept: application/json
    ```

    ## Resources

    The term `resource` refers to a single type of object in the Airflow metadata. An API is broken up by its
    endpoint's corresponding resource.
    The name of a resource is typically plural and expressed in camelCase. Example: `dagRuns`.

    Resource names are used as part of endpoint URLs, as well as in API parameters and responses.

    ## CRUD Operations

    The platform supports **C**reate, **R**ead, **U**pdate, and **D**elete operations on most resources.
    You can review the standards for these operations and their standard parameters below.

    Some endpoints have special behavior as exceptions.

    ### Create

    To create a resource, you typically submit an HTTP `POST` request with the resource's required metadata
    in the request body.
    The response returns a `201 Created` response code upon success with the resource's metadata, including
    its internal `id`, in the response body.

    ### Read

...

...

...

...
    ```
...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /dags/{dag_id}:
    parameters:
      - $ref: "#/components/parameters/DAGID"

    get:
      summary: Get basic information about a DAG
      description: >
        Presents only information available in database (DAGModel).

        If you need detailed information, consider using GET /dags/{dag_id}/details.
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: get_dag
      tags: [DAG]
      parameters:
        - $ref: "#/components/parameters/ReturnFields"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DAG"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update a DAG
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: patch_dag
      parameters:
        - $ref: "#/components/parameters/UpdateMask"
      tags: [DAG]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DAG"
            example:
              is_paused: true
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DAG"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete a DAG
      description: |
        Deletes all metadata related to the DAG, including finished DAG Runs and Tasks.
        Logs are not deleted. This action cannot be undone.

        *New in version 2.2.0*
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_endpoint
      operationId: delete_dag
      tags: [DAG]
      responses:
        "204":
          description: Success.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/AlreadyExists"
...
  /dags/{dag_id}/dagRuns:
    parameters:
      - $ref: "#/components/parameters/DAGID"

    get:
      summary: List DAG runs
      description: >
        This endpoint allows specifying `~` as the dag_id to retrieve DAG runs for all DAGs.
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_run_endpoint
      operationId: get_dag_runs
      tags: [DAGRun]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageOffset"
        - $ref: "#/components/parameters/FilterExecutionDateGTE"
        - $ref: "#/components/parameters/FilterExecutionDateLTE"
        - $ref: "#/components/parameters/FilterStartDateGTE"
        - $ref: "#/components/parameters/FilterStartDateLTE"
        - $ref: "#/components/parameters/FilterEndDateGTE"
        - $ref: "#/components/parameters/FilterEndDateLTE"
        - $ref: "#/components/parameters/FilterUpdatedAtGTE"
        - $ref: "#/components/parameters/FilterUpdatedAtLTE"
        - $ref: "#/components/parameters/FilterState"
        - $ref: "#/components/parameters/OrderBy"
        - $ref: "#/components/parameters/ReturnFields"
      responses:
        "200":
          description: List of DAG runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DAGRunCollection"
        "401":
          $ref: "#/components/responses/Unauthenticated"

    post:
      summary: Trigger a new DAG run.
      description: >
        This will initiate a dagrun.
        If DAG is paused then dagrun state will remain queued, and the task won't run.
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_run_endpoint
      operationId: post_dag_run
      tags: [DAGRun]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DAGRun"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DAGRun"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/AlreadyExists"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
  /dags/~/dagRuns/list:
    post:
      summary: List DAG runs (batch)
      description: >
        This endpoint is a POST to allow filtering across a large number of DAG IDs, where as a GET it
        would run in to maximum HTTP request URL length limit.
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_run_endpoint
      operationId: get_dag_runs_batch
      tags: [DAGRun]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListDagRunsForm"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DAGRunCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /dags/{dag_id}/dagRuns/{dag_run_id}:
    parameters:
      - $ref: "#/components/parameters/DAGID"
      - $ref: "#/components/parameters/DAGRunID"

    get:
      summary: Get a DAG run
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_run_endpoint
      operationId: get_dag_run
      tags: [DAGRun]
      parameters:
        - $ref: "#/components/parameters/ReturnFields"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DAGRun"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete a DAG run
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_run_endpoint
      operationId: delete_dag_run
      tags: [DAGRun]
      responses:
        "204":
          description: Success.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Modify a DAG run
      description: |
        Modify a DAG run.

        *New in version 2.2.0*
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_run_endpoint
      operationId: update_dag_run_state
      tags: [DAGRun]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDagRunState"

      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DAGRun"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
  /dags/{dag_id}/dagRuns/{dag_run_id}/clear:
    parameters:
      - $ref: "#/components/parameters/DAGID"
      - $ref: "#/components/parameters/DAGRunID"

    post:
      summary: Clear a DAG run
      description: |
        Clear a DAG run.

        *New in version 2.4.0*
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_run_endpoint
      operationId: clear_dag_run
      tags: [DAGRun]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClearDagRun"

      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/DAGRun"
                  - $ref: "#/components/schemas/TaskInstanceCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
  /dags/{dag_id}/dagRuns/{dag_run_id}/upstreamAssetEvents:
    parameters:
      - $ref: "#/components/parameters/DAGID"
      - $ref: "#/components/parameters/DAGRunID"
    get:
      summary: Get asset events for a DAG run
      description: |
        Get asset for a dag run.

        *New in version 2.4.0*
        *Changed in 3.0.0*: The endpoint value was renamed from "/dags/{dag_id}/dagRuns/{dag_run_id}/upstreamDatasetEvents"
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_run_endpoint
      operationId: get_upstream_asset_events
      tags: [DAGRun, Asset]
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetEventCollection"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
  /dags/{dag_id}/dagRuns/{dag_run_id}/setNote:
    parameters:
      - $ref: "#/components/parameters/DAGID"
      - $ref: "#/components/parameters/DAGRunID"
    patch:
      summary: Update the DagRun note.
      description: |
        Update the manual user note of a DagRun.

        *New in version 2.5.0*
      x-openapi-router-controller: airflow.api_connexion.endpoints.dag_run_endpoint
      operationId: set_dag_run_note
      tags: [DAGRun]
      requestBody:
        description: Parameters of set DagRun note.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDagRunNote"

      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DAGRun"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  parameters:
...
    PageOffset:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
      description: The number of items to skip before starting to collect the result set.
    PageLimit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        default: 100
      description: The numbers of items to return.

    # Database entity fields
...
    DAGID:
      in: path
      name: dag_id
      schema:
        type: string
      required: true
      description: The DAG ID.
...
    FilterExecutionDateGTE:
      in: query
      name: execution_date_gte
      schema:
        type: string
        format: date-time
      required: false
      description: |
        Returns objects greater or equal to the specified date.

        This can be combined with execution_date_lte parameter to receive only the selected period.
    FilterExecutionDateLTE:
      in: query
      name: execution_date_lte
      schema:
        type: string
        format: date-time
      required: false
      description: |
        Returns objects less than or equal to the specified date.

        This can be combined with execution_date_gte parameter to receive only the selected period.
    FilterStartDateGTE:
      in: query
      name: start_date_gte
      schema:
        type: string
        format: date-time
      required: false
      description: |
        Returns objects greater or equal the specified date.

        This can be combined with start_date_lte parameter to receive only the selected period.
    FilterStartDateLTE:
      in: query
      name: start_date_lte
      schema:
        type: string
        format: date-time
      required: false
      description: |
        Returns objects less or equal the specified date.

        This can be combined with start_date_gte parameter to receive only the selected period.
    FilterEndDateGTE:
      in: query
      name: end_date_gte
      schema:
        type: string
        format: date-time
      required: false
      description: |
        Returns objects greater or equal the specified date.

        This can be combined with start_date_lte parameter to receive only the selected period.
    FilterEndDateLTE:
      in: query
      name: end_date_lte
      schema:
        type: string
        format: date-time
      required: false
      description: |
        Returns objects less than or equal to the specified date.

        This can be combined with start_date_gte parameter to receive only the selected period.
...
    FilterState:
      in: query
      name: state
      schema:
        type: array
        items:
          type: string
      required: false
      description: The value can be repeated to retrieve multiple matching values (OR condition).
...
    FilterUpdatedAtLTE:
      in: query
      name: updated_at_lte
      schema:
        type: string
        format: date-time
      required: false
      description: |
        Returns objects less or equal the specified date.

        This can be combined with updated_at_gte parameter to receive only the selected period.

        *New in version 2.6.0*
    FilterUpdatedAtGTE:
      in: query
      name: updated_at_gte
      schema:
        type: string
        format: date-time
      required: false
      description: |
        Returns objects greater or equal the specified date.

        This can be combined with updated_at_lte parameter to receive only the selected period.

        *New in version 2.6.0*
...
    UpdateMask:
      in: query
      name: update_mask
      schema:
        type: array
        items:
          type: string
      description: |
        The fields to update on the resource. If absent or empty, all modifiable fields are updated.
        A comma-separated list of fully qualified names of fields.
      style: form
      explode: false
    ReturnFields:
      in: query
      name: fields
      schema:
        type: array
        items:
          type: string
      description: |
        List of field for return.

  # Reusable request bodies
...
  responses:
...
    "BadRequest":
      description: Client specified an invalid argument.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # 401
    "Unauthenticated":
      description: Request not authenticated due to missing, invalid, authentication info.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # 403
    "PermissionDenied":
      description: Client does not have sufficient permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # 404
    "NotFound":
      description: A specified resource is not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # 405
...
    "AlreadyExists":
      description: An existing resource conflicts with the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # 500
...
  securitySchemes:
    Basic:
      type: http
      scheme: basic
    GoogleOpenId:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
    Kerberos:
      type: http
      scheme: negotiate

# The API will provide support for plugins to support various authorization mechanisms.
# Detailed information will be available in the plugin specification.
security: []
...